+++ tests/TestCase/View/Helper/ThumbHelperTest.php
@@ -161,6 +161,35 @@ class ThumbHelperTest extends TestCase
     }
 
     /**
+     * Test for `resizeCanvas()` and `resizeCanvasUrl()` methods
+     * @test
+     */
+    public function testResizeCanvas()
+    {
+        $url = $this->Thumb->resizeCanvasUrl('400x400.png', ['width' => 200]);
+        $this->assertRegExp('/^http:\/\/localhost\/thumb\/[A-z0-9]+/', $url);
+
+        $html = $this->Thumb->resizeCanvas('400x400.png', ['width' => 200]);
+        $expected = ['img' => ['src' => $url, 'alt' => '']];
+        $this->assertHtml($expected, $html);
+    }
+
+    /**
+     * Test for `resizeCanvas()` and `resizeCanvasUrl()` methods, with the
+     *  `fullBase` option as `false`
+     * @test
+     */
+    public function testResizeCanvasFullBaseFalse()
+    {
+        $url = $this->Thumb->resizeCanvasUrl('400x400.png', ['width' => 200], ['fullBase' => false]);
+        $this->assertRegExp('/^\/thumb\/[A-z0-9]+/', $url);
+
+        $html = $this->Thumb->resizeCanvas('400x400.png', ['width' => 200], ['fullBase' => false]);
+        $expected = ['img' => ['src' => $url, 'alt' => '']];
+        $this->assertHtml($expected, $html);
+    }
+
+    /**
      * Test for `url` option
      * @return void
      * @test