+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.x branch
 ## 1.3 branch
 # 1.3.1
-* added `ThumbCreator::resizeCanvas()` method.
+* added `ThumbCreator::resizeCanvas()` method;
+* added `ThumbHelper::resizeCanvas()` and `ThumbHelper::resizeCanvasUrl()` methods. 
 
 # 1.3.0
 * the name of the thumbnails (when automatically generated) now contains a